requirements.txt
requests
click
python-dotenv
tqdm
